adiciona inclusao de pedido

+++ src/main/java/br/com/lojavirtual/model/Pedido.java
@@ -24,7 +24,7 @@ public class Pedido {
 
     public void calcularTotal() {
         this.total = itens.stream()
-                .mapToDouble(item -> item.getPrecoUnitario() * item.getQuantidade())
+                .mapToDouble(item -> item.getPrecoUnitario().doubleValue() * item.getQuantidade())
                 .sum();
     }
 }
